[refactor] Change Signup to Register
